+++ frontend/package.json
@@ -11,8 +11,10 @@
     "test": "jest"
   },
   "dependencies": {
+    "@codemirror/commands": "^6.1.2",
     "@codemirror/lang-javascript": "^6.1.1",
     "@codemirror/state": "^6.1.4",
+    "@codemirror/view": "^6.6.0",
     "codemirror": "^6.0.1",
     "jest-dom": "^4.0.0",
     "lib0": "^0.2.54",
